More formatting changes to misc_actor code in default missionpack and Lazarus DLLs. Added support for LMSP plasma rifle to misc_actor in missionpack DLL.

+++ game/m_actor_weap.c
@@ -59,7 +59,7 @@ void ActorTarget (edict_t *self, vec3_t target)
 	vec3_t	v;
 
 	if (!self->enemy) {
-		VectorClear(target);
+		VectorClear (target);
 		return;
 	}
 	if (self->monsterinfo.aiflags & AI_GOOD_GUY)
@@ -69,9 +69,9 @@ void ActorTarget (edict_t *self, vec3_t target)
 
 	if (self->enemy->health > 0)
 	{
-		int		weapon;
-		trace_t	tr;
-		vec3_t	start;
+		int			weapon;
+		trace_t		tr;
+		vec3_t		start;
 		qboolean	can_see=false;
 
 		VectorCopy(self->s.origin,start);
@@ -146,7 +146,7 @@ void actorBlaster (edict_t *self)
 #endif
 	}
 #ifdef KMQUAKE2_ENGINE_MOD
-	else if (sk_blaster_color->value == 4) {	// red
+	else if ((int)sk_blaster_color->value == 4) {	// red
 		color = BLASTER_RED;
 		effect = EF_BLASTER|EF_IONRIPPER;
 	}
@@ -478,37 +478,38 @@ void actorGrenadeLauncher (edict_t *self)
 		float	drop;
 		float	last_error, last_up, v_error;
 		int		i;
+
 		VectorCopy (forward, target);	// save target point
 		// horizontal distance to target
-		x = sqrt( forward[0]*forward[0] + forward[1]*forward[1]);
-		cosa = sqrt(aim[0]*aim[0] + aim[1]*aim[1]);
+		x = sqrt( forward[0] * forward[0] + forward[1] * forward[1]);
+		cosa = sqrt(aim[0] * aim[0] + aim[1] * aim[1]);
 		// constant horizontal velocity (since grenades don't have drag)
 		vx = GRENADE_VELOCITY * cosa;
 		// time to reach target x
-		t = x/vx;
+		t = x / vx;
 		// in that time, grenade will drop this much:
-		drop = 0.5*sv_gravity->value*t*(t+FRAMETIME);
+		drop = 0.5 * sv_gravity->value * t * (t + FRAMETIME);
 		forward[2] = target[2] + drop;
 		// this is a good first cut, but incorrect since angle now changes, so
 		// horizontal speed changes
 		VectorCopy (forward, aim);
 		VectorNormalize (aim);
-		cosa = sqrt(aim[0]*aim[0] + aim[1]*aim[1]);
+		cosa = sqrt(aim[0] * aim[0] + aim[1] * aim[1]);
 		vx = GRENADE_VELOCITY * cosa;
-		t = x/vx;
-		y = GRENADE_VELOCITY*aim[2]*t - 0.5*sv_gravity->value*t*(t+FRAMETIME);
-		v_error = target[2]-y;
-		last_error = 2*v_error;
+		t = x / vx;
+		y = GRENADE_VELOCITY * aim[2] * t - 0.5 * sv_gravity->value * t * (t + FRAMETIME);
+		v_error = target[2] - y;
+		last_error = 2 * v_error;
 		for (i=0; i<10 && fabs(v_error) > 4 && fabs(v_error) < fabs(last_error); i++)
 		{
-			drop = 0.5*sv_gravity->value*t*(t+FRAMETIME);
+			drop = 0.5 * sv_gravity->value * t * (t + FRAMETIME);
 			forward[2] = target[2] + drop;
 			VectorCopy (forward, aim);
 			VectorNormalize (aim);
-			cosa = sqrt(aim[0]*aim[0] + aim[1]*aim[1]);
+			cosa = sqrt(aim[0] * aim[0] + aim[1] * aim[1]);
 			vx = GRENADE_VELOCITY * cosa;
-			t = x/vx;
-			y = GRENADE_VELOCITY*aim[2]*t - 0.5*sv_gravity->value*t*(t+FRAMETIME);
+			t = x / vx;
+			y = GRENADE_VELOCITY * aim[2] * t - 0.5 * sv_gravity->value * t * (t + FRAMETIME);
 			v_error = target[2]-y;
 			// If error is increasing... we can't get there from here and
 			// probably shouldn't be here in the first place. Too late now...
@@ -536,13 +537,13 @@ void actorGrenadeLauncher (edict_t *self)
 				// OK... the aim vector hit a solid, but would the grenade actually hit?
 				int		contents;
 
-				cosa = sqrt(aim[0]*aim[0] + aim[1]*aim[1]);
+				cosa = sqrt(aim[0] * aim[0] + aim[1] * aim[1]);
 				vx = GRENADE_VELOCITY * cosa;
 				VectorSubtract (tr.endpos, start, dist);
 				dist[2] = 0;
 				x = VectorLength(dist);
-				t = x/vx;
-				drop = 0.5*sv_gravity->value*t*(t+FRAMETIME);
+				t = x / vx;
+				drop = 0.5 * sv_gravity->value * t * (t + FRAMETIME);
 				tr.endpos[2] -= drop;
 				// move just a bit in the aim direction
 				tr.endpos[0] += aim[0];
@@ -557,13 +558,13 @@ void actorGrenadeLauncher (edict_t *self)
 					tr = gi.trace(start, vec3_origin, vec3_origin, target, self, MASK_SOLID);
 					if (tr.fraction < 1.0)
 					{
-						cosa = sqrt(aim[0]*aim[0] + aim[1]*aim[1]);
+						cosa = sqrt(aim[0] * aim[0] + aim[1] * aim[1]);
 						vx = GRENADE_VELOCITY * cosa;
 						VectorSubtract (tr.endpos, start, dist);
 						dist[2] = 0;
 						x = VectorLength(dist);
-						t = x/vx;
-						drop = 0.5*sv_gravity->value*t*(t+FRAMETIME);
+						t = x / vx;
+						drop = 0.5 * sv_gravity->value * t * (t + FRAMETIME);
 						tr.endpos[2] -= drop;
 						tr.endpos[0] += aim[0];
 						tr.endpos[1] += aim[1];
@@ -587,7 +588,7 @@ void actorGrenadeLauncher (edict_t *self)
 
 		VectorCopy (self->velocity, v1);
 		VectorNormalize (v1);
-		delta = -monster_speed/GRENADE_VELOCITY;
+		delta = -monster_speed / GRENADE_VELOCITY;
 		VectorMA (aim, delta, v1, aim);
 		VectorNormalize (aim);
 	}
@@ -665,12 +666,12 @@ void actorHyperblaster (edict_t *self)
 	else
 	{
 		// Knightmare- select color
-		if (sk_hyperblaster_color->value == 2)		// green
+		if ((int)sk_hyperblaster_color->value == 2)		// green
 			color = BLASTER_GREEN;
-		else if (sk_hyperblaster_color->value == 3)	// blue
+		else if ((int)sk_hyperblaster_color->value == 3)	// blue
 			color = BLASTER_BLUE;
 	#ifdef KMQUAKE2_ENGINE_MOD
-		else if (sk_hyperblaster_color->value == 4)	// red
+		else if ((int)sk_hyperblaster_color->value == 4)	// red
 			color = BLASTER_RED;
 	#endif
 		else										// standard yellow
@@ -683,12 +684,12 @@ void actorHyperblaster (edict_t *self)
 		VectorNormalize (forward);
 		if ((random() * 3) < 1)
 		{
-			if (sk_hyperblaster_color->value == 2)		// green
+			if ((int)sk_hyperblaster_color->value == 2)		// green
 				effect = (EF_HYPERBLASTER|EF_TRACKER);
-			else if (sk_hyperblaster_color->value == 3)	// blue
+			else if ((int)sk_hyperblaster_color->value == 3)	// blue
 				effect = EF_BLUEHYPERBLASTER;
 	#ifdef KMQUAKE2_ENGINE_MOD
-			else if (sk_hyperblaster_color->value == 4)	// red
+			else if ((int)sk_hyperblaster_color->value == 4)	// red
 				effect = EF_HYPERBLASTER|EF_IONRIPPER;
 	#endif
 			else										// standard yellow
